Migrated all classes to a new Class Library project for future expansion of the project to include diffrent .Net features without redundancy problems.
