+++ ConsoleCalculator.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34408.163
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleCalculator", "ConsoleCalculator\ConsoleCalculator.csproj", "{3FEDA428-BA5E-4783-B369-E39D222ADACE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleCalculator", "ConsoleCalculator\ConsoleCalculator.csproj", "{3FEDA428-BA5E-4783-B369-E39D222ADACE}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator", "Calculator\Calculator.csproj", "{8B595842-E33C-4355-8A6E-C9684B0D4EE0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{3FEDA428-BA5E-4783-B369-E39D222ADACE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3FEDA428-BA5E-4783-B369-E39D222ADACE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3FEDA428-BA5E-4783-B369-E39D222ADACE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8B595842-E33C-4355-8A6E-C9684B0D4EE0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8B595842-E33C-4355-8A6E-C9684B0D4EE0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8B595842-E33C-4355-8A6E-C9684B0D4EE0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8B595842-E33C-4355-8A6E-C9684B0D4EE0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
